pyproject.toml
[tool.poetry]
name = "auth-app"
version = "0.1.0"
description = ""
authors = ["OLuwatobiTobias <giwaoluwatobi@gmail.com>"]
readme = "README.md"
packages = [{include = "auth_app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.91.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
